add the function of reading 'out of stocks' from the database and display it in the right table
server.js - change the addition to the database for 'out of stock' and
add the functions of finding the right information from
it.
outofstock.html - remove the manual data and use angular to insert
the data from the database.
managers.html - using the 'loadStockDB' for
loading the data from the database.
managersapp.js - adding the
functions that handle the pulling of the right data from the
database.
outofstock.js - change some attributes.