add the shift and baking cakes requests functionality
add the functionality to request a shift for volunteers.
add the functionality to add a baking cake request for bakers.
add the ability to see the requests for managers.
server.js - saving the requests to the DB(for both, bakers and
volunteers) and reload from it(for managers).
style.css - add the request system style to css file.
models - add few models - cakes, cakeRequests, shifts, shiftRequests